part_4 exercise_4.21

+++ controllers/blogs.js
@@ -52,8 +52,17 @@ blogsRouter.put('/:id', async (req, res) => {
 
 blogsRouter.delete('/:id', async (req, res, next) => {
   try {
-    await Blog.findByIdAndRemove(req.params.id);
-    res.status(204).end();
+    const blog = await Blog.findById(req.params.id);
+    const decodedToken = jwt.verify(req.token, process.env.SECRET);
+    if (!decodedToken.id) {
+      return res.status(401).json({ error: 'token missing or invalid' });
+    }
+    const user = await User.findById(decodedToken.id);
+
+    if (blog && user && blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndRemove(req.params.id);
+      res.status(204).end();
+    }
   } catch (e) {
     next(e);
   }
